+++ pipeline/cache/engine.go
@@ -13,7 +13,6 @@ import (
 	"go.uber.org/zap"
 
 	pbsubstreams "github.com/streamingfast/substreams/pb/sf/substreams/v1"
-	"github.com/streamingfast/substreams/service/config"
 	"github.com/streamingfast/substreams/storage/execout"
 )
 
@@ -30,14 +29,12 @@ type Engine struct {
 	execOutputWriters map[string]*execout.Writer // moduleName => writer (single file)
 	existingExecOuts  map[string]*execout.File
 
-	runtimeConfig config.RuntimeConfig // TODO(abourget): Deprecated: remove this as it's not used
-	logger        *zap.Logger
+	logger *zap.Logger
 }
 
-func NewEngine(ctx context.Context, runtimeConfig config.RuntimeConfig, execOutWriters map[string]*execout.Writer, blockType string, existingExecOuts map[string]*execout.File) (*Engine, error) {
+func NewEngine(ctx context.Context, execOutWriters map[string]*execout.Writer, blockType string, existingExecOuts map[string]*execout.File) (*Engine, error) {
 	e := &Engine{
 		ctx:               ctx,
-		runtimeConfig:     runtimeConfig,
 		reversibleBuffers: map[uint64]*execout.Buffer{},
 		execOutputWriters: execOutWriters,
 		logger:            reqctx.Logger(ctx),