+++ docs/release-notes/change-log.md
@@ -9,6 +9,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.5.6 (Unreleased)
+
+### Fixes
+
+* Fix bug where substreams tier2 would sometimes write outputs with the wrong tag (leaked from another tier1 request)
+
+### Remove
+
+* Removed MaxWasmFuel since it is not supported in Wazero
+
 ## v1.5.5
 
 ### Fixes